worker/requirements.txt
pika
redis
requests
ffmpeg-python
